+++ pdf/lib/widgets.dart
@@ -28,6 +28,7 @@ export 'src/widgets/chart/grid_cartesian.dart';
 export 'src/widgets/chart/grid_radial.dart';
 export 'src/widgets/chart/legend.dart';
 export 'src/widgets/chart/line_chart.dart';
+export 'src/widgets/chart/pie_chart.dart';
 export 'src/widgets/clip.dart';
 export 'src/widgets/container.dart';
 export 'src/widgets/content.dart';